Extract llvm datalayout parsing out of spec module
